LayoutTests/editing/selection/ios/change-selection-after-tapping-focused-element-expected.txt

WebKit
Verifies that after tapping a focused element to bring up the keyboard, changing the selection via script causes the native selection UI to be shown. To manually test, tap the red box above and tap select all via the callout bar; the entire text in the editable element should be selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selectionRects.length is 1
PASS selectionRects[0].left is 2
PASS selectionRects[0].top is 2
PASS selectionRects[0].width is 309
PASS selectionRects[0].height is 114
PASS successfullyParsed is true

TEST COMPLETE
